+++ src/main/java/nextstep/app/application/MemberService.java
@@ -3,7 +3,7 @@ package nextstep.app.application;
 import nextstep.app.domain.MemberRepository;
 import nextstep.app.payload.MemberDetails;
 import nextstep.app.ui.AuthenticationException;
-import nextstep.security.service.UserDetailsService;
+import nextstep.security.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
 @Service
