+++ src/main/java/nextstep/security/authentication/DaoAuthenticationProvider.java
@@ -1,7 +1,7 @@
 package nextstep.security.authentication;
 
 import nextstep.app.ui.AuthenticationException;
-import nextstep.security.service.UserDetailsService;
+import nextstep.security.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
 @Service
