+++ src/main/java/nextstep/security/userdetails/UserDetailsService.java
@@ -1,4 +1,4 @@
-package nextstep.security.service;
+package nextstep.security.userdetails;
 
 public interface UserDetailsService {
     UserDetails findUserDetailsByPrincipal(Object principal);
